email_service/requirements.txt
grpcio-tools==1.21.1
